apps/openclaw-listen/pyproject.toml
[project]
name = "openclaw-listen"
version = "0.1.0"
description = "OpenClaw-native async job listener for linux-agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.28",
  "pyyaml>=6.0",
  "pydantic>=2.6",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main", "worker"]
